update to move exception formatting logic to RequestClassifier
