+++ sdk/core/Azure.Core/src/Pipeline/HttpPipeline.cs
@@ -26,15 +26,9 @@ namespace Azure.Core.Pipeline
         /// <param name="policies">Policies to be invoked as part of the pipeline in order.</param>
         /// <param name="responseClassifier">The response classifier to be used in invocations.</param>
         public HttpPipeline(HttpPipelineTransport transport, HttpPipelinePolicy[]? policies = null, ResponseClassifier? responseClassifier = null)
-            : this(transport, policies, responseClassifier, null)
-        {
-        }
-
-        internal HttpPipeline(HttpPipelineTransport transport, HttpPipelinePolicy[]? policies = null, ResponseClassifier? responseClassifier = null, ResponseExceptionFactory? exceptionFactory = null)
         {
             _transport = transport ?? throw new ArgumentNullException(nameof(transport));
             ResponseClassifier = responseClassifier ?? new ResponseClassifier();
-            ExceptionFactory = exceptionFactory ?? new ResponseExceptionFactory(new DefaultClientOptions());
 
             policies = policies ?? Array.Empty<HttpPipelinePolicy>();
 
@@ -66,8 +60,6 @@ namespace Azure.Core.Pipeline
         /// </summary>
         public ResponseClassifier ResponseClassifier { get; }
 
-        internal ResponseExceptionFactory ExceptionFactory { get; }
-
         /// <summary>
         /// Invokes the pipeline asynchronously. After the task completes response would be set to the <see cref="HttpMessage.Response"/> property.
         /// </summary>
@@ -80,7 +72,6 @@ namespace Azure.Core.Pipeline
             AddHttpMessageProperties(message);
             var value = _pipeline.Span[0].ProcessAsync(message, _pipeline.Slice(1));
             message.Response.ResponseClassifier = ResponseClassifier;
-            message.Response.ExceptionFactory = ExceptionFactory;
             return value;
         }
 
@@ -95,7 +86,6 @@ namespace Azure.Core.Pipeline
             AddHttpMessageProperties(message);
             _pipeline.Span[0].Process(message, _pipeline.Slice(1));
             message.Response.ResponseClassifier = ResponseClassifier;
-            message.Response.ExceptionFactory = ExceptionFactory;
         }
         /// <summary>
         /// Invokes the pipeline asynchronously with the provided request.
